Crate scryer_prolog

A free software ISO Prolog system.

Structs

Machine

An instance of Scryer Prolog.

MachineConfig

Describes how a Machine will be configured.

QueryState

An iterator though the leaf answers of a query.

StreamConfig

Describes how the streams of a Machine will be handled.

Enums

LeafAnswer

Represents a leaf answer from a query.

PrologTerm

Represents a Prolog term.

Functions

default_toplevel

Returns a static string slice to the default toplevel

run_binary

The entry point for the Scryer Prolog CLI.